Fully working GEDCOM Parsing
Parses the GEDCOM into two seperate .csv files - one for families and one for individuals and then in the other file it runs the .csv files through prettytable to get some pretty tables. Might just have the beginning of the main file call the gedcom2csv to start instead of running one file then the other.